fix(constants): Export RasterBandColorinterp
